Change the format used to store log files on disk.
This allows packages to be linked to the sha1 sum of their ebuilds to
prevent mixed information when ebuilds are edited but not bumped. It
also lays the groundwork for including information about hypothetical
build situations, next_target, next_keyword, ect.

The db schema, web interface, and docker glue will need to be updated to
account for this new layout.